+++ lib/infra/storage/app_cache_manager.dart
@@ -1,0 +1,6 @@
+import 'package:revelation/infra/storage/app_cache_manager_stub.dart'
+    if (dart.library.io) 'package:revelation/infra/storage/app_cache_manager_io.dart'
+    if (dart.library.js_interop) 'package:revelation/infra/storage/app_cache_manager_web.dart'
+    as impl;
+
+Future<void> clearAppCache() => impl.clearAppCache();